reduce mtu to 1394 on supported client
Clients sending ethernet frames with a payload larger than 1394 will cause
a fragmentation of the ethernet frame. This increases the overhead of
transmitting it and increases the chance that this frame is lost.

Letting supported clients know about the reduced mtu on the server side
reduces the amount of fragmented packets.