Replace nodes.json.priv with atomic operation and only when valid
The nodes.json.priv is generated using jq which does this using multiple
write. So a different program (webserver) might try to read the half
written file and thus the meshviewer client will fail to correctly read it.

Instead replace the old file using a simple move. And also make sure that
the new file is valid json to avoid problems when input nodes.json was
already broken.